+++ src/knnresult.jl
@@ -1,6 +1,6 @@
 # This file is a part of SimilaritySearch.jl
 export KnnResult
-export maxlength, maxlength, getpair, getdist, getid, initialstate, idview, distview, reuse!
+export maxlength, maxlength, getdist, getid, idview, distview, reuse!
 
 """
     KnnResult(ksearch::Integer)
@@ -119,9 +119,8 @@ Returns a result set and a new initial state; reuse the memory buffers
 """
 @inline function reuse!(res::KnnResult, k::Integer=res.k)
     @assert k > 0
-    empty!(res.id)
-    empty!(res.dist)
-
+    empty!(res.id); empty!(res.dist)
+    # sizehint!(res.id, k), sizehint!(res.dist, k)
     KnnResult(res.id, res.dist, k)
 end
 
