+++ joss/paper.bib
@@ -24,7 +24,8 @@
   number={11},
   pages={2278--2324},
   year={1998},
-  publisher={Ieee}
+  publisher={Ieee},
+  doi={10.1109/5.726791}
 }
 @inproceedings{nndescent11,
 author = {Dong, Wei and Moses, Charikar and Li, Kai},
@@ -33,7 +34,6 @@ year = {2011},
 isbn = {9781450306324},
 publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
--url = {https://doi.org/10.1145/1963405.1963487},
 doi = {10.1145/1963405.1963487},
 booktitle = {Proceedings of the 20th International Conference on World Wide Web},
 pages = {577--586},
@@ -56,7 +56,8 @@ series = {WWW '11}
   author={Rubinstein, Aviad},
   booktitle={Proceedings of the 50th annual ACM SIGACT symposium on theory of computing},
   pages={1260--1268},
-  year={2018}
+  year={2018},
+  doi={10.1145/3188745.3188916}
 }
 
 @article{sparse-dense-text-retrieval,
@@ -125,7 +126,8 @@ series = {WWW '11}
   number={5500},
   pages={2319--2323},
   year={2000},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.290.5500.2319}
 }
 
 
@@ -152,7 +154,8 @@ series = {WWW '11}
   booktitle={International Conference on Similarity Search and Applications},
   pages={297--311},
   year={2020},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-60936-8_23}
 }
 
 @article{jayaram2019diskann,
@@ -170,7 +173,8 @@ series = {WWW '11}
   number={2},
   pages={763--777},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10044-020-00946-w}
 }
 
 @inproceedings{ruiz2015finding,
@@ -179,7 +183,8 @@ series = {WWW '11}
   booktitle={International Conference on Similarity Search and Applications},
   pages={103--109},
   year={2015},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-25087-8_10}
 }
 
 @article{malkov2018efficient,
@@ -190,7 +195,8 @@ series = {WWW '11}
   number={4},
   pages={824--836},
   year={2018},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tpami.2018.2889473}
 }
 
 @article{malkov2014approximate,
@@ -200,7 +206,8 @@ series = {WWW '11}
   volume={45},
   pages={61--68},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.is.2013.10.006}
 }
 
 @inproceedings{chavez2010navigating,
@@ -209,7 +216,8 @@ series = {WWW '11}
   booktitle={Mexican Conference on Pattern Recognition},
   pages={270--280},
   year={2010},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-642-15992-3_29}
 }
 
 @article{Tellez2013,
@@ -220,17 +228,10 @@ number = {7},
 pages = {1019-1030},
 year = {2013},
 issn = {0306-4379},
-doi = {https://doi.org/10.1016/j.is.2012.06.005},
+doi = {10.1016/j.is.2012.06.005},
[…]
