### 2022-05-26 #4421 @sadit
Updated paper.md, fixed missing DOIs
