+++ joss/paper.bib
@@ -43,12 +43,19 @@ location = {Hyderabad, India},
 series = {WWW '11}
 }
 
+
 @inproceedings{scann2020,
-  title={Accelerating Large-Scale Inference with Anisotropic Vector Quantization},
   author={Guo, Ruiqi and Sun, Philip and Lindgren, Erik and Geng, Quan and Simcha, David and Chern, Felix and Kumar, Sanjiv},
-  booktitle={International Conference on Machine Learning},
-  year={2020},
-  -URL={https://arxiv.org/abs/1908.10396}
+  booktitle={37th International Conference on Machine Learning, ICML 2020},
+  -URL={https://arxiv.org/abs/1908.10396},
+	title = {Accelerating large-scale inference with anisotropic vector quantization},
+	year = {2020},
+	volume = {PartF168147-5},
+	pages = {3845--3854},
+	url = {https://www.scopus.com/inward/record.uri?eid=2-s2.0-85105244400&partnerID=40&md5=e3ab797435367141112b5e5843b2cb1e},
+	publication_stage = {Final},
+	note = {Cited by: 34}
+}
 }
 
 @inproceedings{rubinstein2018hardness,
@@ -460,15 +467,6 @@ doi={10.1007/978-3-319-25087-8_10}
   HAL_VERSION = {v2},
 }
 
-@inproceedings{hoyos2017self,
-  title={Self-indexed Motion Planning},
-  author={Hoyos, Angello and Ruiz, Ubaldo and Tellez, Eric and Chavez, Edgar},
-  booktitle={International Conference on Similarity Search and Applications},
-  pages={220--233},
-  year={2017},
-  organization={Springer}
-}
-
 @inproceedings{sash,
 	Author = {Houle, Michael E and Sakuma, Jun},
 	Booktitle = {Data Engineering, 2005. ICDE 2005. Proceedings. 21st International Conference on},
@@ -681,21 +679,6 @@ doi={10.1007/978-3-319-25087-8_10}
 	Volume = {45},
 	Year = {2014},
 	Bdsk-Url-1 = {http://dx.doi.org/10.1016/j.is.2013.10.006}}
-
-@article{is2014knr,
-	Author = {Edgar Ch{\'a}vez and Mario Graff and Gonzalo Navarro and Eric S. T{\'e}llez},
-	-Doi = {10.1016/j.is.2015.02.001},
-	Issn = {0306-4379},
-	Journal = {Information Systems},
-	Keywords = {Indexing metric spaces},
-	Number = {To be determined},
-	Pages = {-},
-	Title = {Near neighbor searching with K nearest references},
-	-Url = {http://www.sciencedirect.com/science/article/pii/S0306437915000241},
-	Volume = {To be determined},
-	Year = {2015, Sent in 2014.},
-	Bdsk-Url-1 = {http://www.sciencedirect.com/science/article/pii/S0306437915000241},
-	Bdsk-Url-2 = {https://dx.doi.org/10.1016/j.is.2015.02.001}}
 
 @book{BurkeKendall2014,
 	Address = {New York, NY, USA},
