### 2022-05-30 #4421 @jbytecode
👋👋👋 Dear @kose-y  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**SimilaritySearch.jl: Autotuned nearest neighbor indexes for Julia**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4421).
