### 2022-05-31 #4442 @editorialbot
[REVIEW]: SimilaritySearch.jl: Autotuned nearest neighbor indexes for Julia
**Submitting author:** @sadit (<a href="http://orcid.org/0000-0001-5804-9868">Eric S. Tellez</a>)
**Repository:** https://github.com/sadit/SimilaritySearch.jl
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.9.3
