### 2022-06-05 4edb608
knnresult set working with pointers
