+++ src/searchgraph/SearchGraph.jl
@@ -6,7 +6,7 @@ export find_neighborhood, push_neighborhood!
 export callbacks
 export BeamSearch, BeamSearchSpace, Callback
 export KDisjointHints, DisjointHints, RandomHints
-
+export RandomPruning, KeepNearestPruning, SatPruning, prune!
 include("graph.jl")
 include("callbacks.jl")
 include("rebuild.jl")
