### 2022-06-07 08f5ad0
ajustments for @batch (set as per=thread, it works better for our benchmarks)
