+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/","page":"API","title":"API","text":"\nCurrentModule = SimilaritySearch\nDocTestSetup = quote\n    using SimilaritySearch\nend","category":"page"},{"location":"api/#Indexes","page":"API","title":"Indexes","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"ExhaustiveSearch\nParallelExhaustiveSearch\nSearchGraph","category":"page"},{"location":"api/#SimilaritySearch.ExhaustiveSearch","page":"API","title":"SimilaritySearch.ExhaustiveSearch","text":"ExhaustiveSearch(dist::SemiMetric, db::AbstractVector)\n\nSolves queries evaluating dist for the query and all elements in the dataset\n\n\n\n\n\n","category":"type"},{"location":"api/#SimilaritySearch.ParallelExhaustiveSearch","page":"API","title":"SimilaritySearch.ParallelExhaustiveSearch","text":"ParallelExhaustiveSearch(; dist=SqL2Distance(), db=VectorDatabase{Float32}())\n\nSolves queries evaluating dist in parallel for the query and all elements in the dataset. Note that this should not be used in conjunction with searchbatch(...; parallel=true) since they will compete for resources.\n\n\n\n\n\n","category":"type"},{"location":"api/#SimilaritySearch.SearchGraph","page":"API","title":"SimilaritySearch.SearchGraph","text":"struct SearchGraph <: AbstractSearchContext\n\nSearchGraph index. It stores a set of points that can be compared through a distance function dist. The performance is determined by the search algorithm search_algo and the neighborhood policy. It supports callbacks to adjust parameters as insertions are made.\n\nhints: Initial points for exploration (empty hints imply using random points)\n\nNote: Parallel insertions should be made through append! or index! function with parallel_block > 1\n\n\n\n\n\n","category":"type"},{"location":"api/#Searching","page":"API","title":"Searching","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"search\nsearchbatch","category":"page"},{"location":"api/#SimilaritySearch.search","page":"API","title":"SimilaritySearch.search","text":"search(seq::ExhaustiveSearch, q, res::KnnResult)\n\nSolves the query evaluating all items in the given query.\n\n\n\n\n\nsearch(ex::ParallelExhaustiveSearch, q, res::KnnResult; minbatch=0, pools=nothing)\n\nSolves the query evaluating all items in the given query.\n\nArguments\n\nex: the search structure\nq: the query to solve\nres: the result set\n\nKeyword arguments\n\nminbatch: Minimum number of queries solved per each thread, see getminibatch\npools: The set of caches (nothing for this index)\n\n\n\n\n\nsearch(bs::BeamSearch, index::SearchGraph, q, res, hints, pools; bsize=bs.bsize, Δ=bs.Δ, maxvisits=bs.maxvisits)\n\nTries to reach the set of nearest neighbors specified in res for q.\n\nbs: the parameters of BeamSearch\nindex: the local search index\nq: the query\nres: The result object, it stores the results and also specifies the kind of query\nhints: Starting points for searching, randomly selected when it is an empty collection\npools: A SearchGraphPools object with preallocated pools\n\nOptional arguments (defaults to values in bs)\n\nbsize: Beam size\nΔ: exploration expansion factor\nmaxvisits: Maximum number of nodes to visit (distance evaluations)\n\n\n\n\n\nsearch(index::SearchGraph, q, res; hints=index.hints, pools=getpools(index))\n\nSolves the specified query res for the query object q.\n\n\n\n\n\n
[…]
