### 2022-06-09 744a01f
allknn benchmark
