+++ benchmarks/allknn.jl
@@ -7,12 +7,21 @@ function create_database()
 end
 
 function main()
+    @info "this benchmark is intended to work with multithreading enabled julia sessions"
     db = create_database()
-    H = SearchGraph(; db, dist=NormalizedCosineDistance(), verbose=false)
+    dist = NormalizedCosineDistance()
+    k = 32
+    @info "----- computing gold standard"
+    goldsearchtime = @elapsed gI, gD = allknn(ExhaustiveSearch(; db, dist), k)
+    @info "----- computing search graph"
+    H = SearchGraph(; db, dist, verbose=false)
     index!(H; parallel_block=128)
-    searchtime = @elapsed hI, hD = allknn(H, 32; parallel=true)
+    optimize!(H, MinRecall(0.9), ksearch=k)
+    searchtime = @elapsed hI, hD = allknn(H, k)
     n = length(db)
-    @show (; n, searchtime, qps=n / searchtime)
+    @info "gold:" (; n, goldsearchtime, qps=n/goldsearchtime)
+    @info "searchgraph:" (; n, searchtime, qps=n / searchtime)
+    @info "recall:" macrorecall(gI, hI)
 end
 
 main()
