+++ Project.toml
@@ -18,10 +18,10 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Distances = "0.10"
 Intersections = "0.2.8"
 Parameters = "0.12"
+Polyester = "0.6"
 SearchModels = "0.3"
 StatsBase = "0.30.0, 0.32.0, 0.33"
 julia = "1.6"
-Polyester = "0.6"
 
 [extras]
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
