+++ src/closestpair.jl
@@ -3,7 +3,7 @@
 export closestpair
 
 """
-    closestpair(idx::AbstractSearchContext; parallel=false, pools=getpools(idx))
+    closestpair(idx::AbstractSearchContext; minbatch=0, pools=getpools(idx))
 
 Finds the closest pair among all elements in `idx`. If the index `idx` is approximate then pair of points could be also an approximation.
 
@@ -11,37 +11,42 @@ Finds the closest pair among all elements in `idx`. If the index `idx` is approx
 - `idx`: the search structure that indexes the set of points
 
 # Keyword Arguments:
-- `parallel`: If true then the algorithm uses all available threads to compute the closest pair
+- `minbatch`: controls how multithreading is used for evaluating configurations, see [`getminbatch`](@ref)
 - `pools`: The pools needed for the index. Only used for special cases, default values should work in most cases. See [`getpools`](@ref) for more information.
 """
-function closestpair(idx::AbstractSearchContext; parallel=false, pools=getpools(idx))
-    parallel ? parallel_closestpair(idx, Threads.nthreads(), pools) : sequential_closestpair(idx, pools)
+function closestpair(idx::AbstractSearchContext; minbatch=0, pools=getpools(idx))
+    if Threads.nthreads() == 1 || minbatch < 0 
+        sequential_closestpair(idx, pools)
+    else
+        parallel_closestpair(idx, pools, minbatch)
+    end
 end
 
 function search_hint(idx::AbstractSearchContext, i::Integer, pools)
-    res = getknnresult(2)
+    res = getknnresult(2, pools)
     search(idx, idx[i], res; pools)
     argmin(res) == i ? (argmax(res), maximum(res)) : (argmin(res), minimum(res))
 end
 
 function search_hint(G::SearchGraph, i::Integer, pools)
-    res = getknnresult(2)
-    search(G.search_algo, G, G[i], res, first(G.links[i]), pools)
-    argmin(res) == i ? (argmax(res), maximum(res)) : (argmin(res), minimum(res))
+    res = getknnresult(3, pools)
+    vstate = getvstate(length(G), pools)
+    visit!(vstate, i)
+    search(G.search_algo, G, G[i], res, rand(G.links[i]), pools; vstate)
+    argmin(res), minimum(res)
 end
 
-function parallel_closestpair(idx::AbstractSearchContext, parallel_block, pools)
+function parallel_closestpair(idx::AbstractSearchContext, pools, minbatch)::Tuple{Int32,Int32,Float32}
     n = length(idx)
-    parallel_block = min(n, parallel_block)
-    B = [(zero(Int32), zero(Int32), typemax(Float32)) for _ in 1:parallel_block]
+    B = [(zero(Int32), zero(Int32), typemax(Float32)) for _ in 1:Threads.nthreads()]
 
-    for block in Iterators.partition(1:n, parallel_block)
-        Threads.@threads for i in 1:length(block)
-            @inbounds objID = block[i]
-            id_, d_ = search_hint(idx, objID, pools)
-            @inbounds if d_ < last(B[i])
-                B[i] = (objID, id_, d_)
-            end
+    minbatch = getminibatch(minbatch, n)
+
+    @batch minbatch=minbatch per=thread for objID in 1:n
+        id_, d_ = search_hint(idx, objID, pools)
+        tID = Threads.threadid()
+        @inbounds if d_ < last(B[tID])
+            B[tID] = (Int32(objID), id_, d_)
         end
     end
 
@@ -49,14 +54,14 @@ function parallel_closestpair(idx::AbstractSearchContext, parallel_block, pools)
     B[i]
 end
 
[…]
