### 2022-06-10 d02193d
renames getminibatch -> getminbatch; some small optimizations; loop vectorization
