### 2022-06-11 d5448b8
fixes some typos in the code; define a default parallel_block for multiprocessing by default
