+++ benchmarks/allknn-lv.jl
@@ -31,7 +31,7 @@ function main()
     index!(H; parallel_block=128)
     optimize!(H, MinRecall(0.9), ksearch=k)
     # prune!(RandomPruning(12), H)
-    searchtime = @elapsed hI, hD = allknn(H, k)
+    searchtime = @elapsed hI, hD = allknn(H, k; minbatch=32)
     n = length(db)
     @info "gold:" (; n, goldsearchtime, qps=n/goldsearchtime)
     @info "searchgraph:" (; n, searchtime, qps=n / searchtime)
