### 2022-06-13 d9718f0
previous to generic KnnResult
