### 2022-06-13 2ffab32
working Generic KnnResult
