+++ src/knnresult.jl
@@ -2,10 +2,54 @@
 export KnnResult
 export maxlength, maxlength, getdist, getid, idview, distview, reuse!
 
-
 struct IdDistArray
     id::Vector{Int32}
     dist::Vector{Float32}
+    k::Int  # number of neighbors
+end
+
+@inline idview(a::IdDistArray) = a.id
+@inline distview(a::IdDistArray) = a.dist
+@inline getid(a::IdDistArray, i::Integer) = @inbounds a.id[i]
+@inline getdist(a::IdDistArray, i::Integer) = @inbounds a.dist[i]
+@inline maxlength(a::IdDistArray) = a.k
+@inline Base.maximum(a::IdDistArray) = last(a.dist)
+@inline Base.argmax(a::IdDistArray) = last(a.id)
+@inline Base.minimum(a::IdDistArray) = first(a.dist)
+@inline Base.argmin(a::IdDistArray) = first(a.id)
+@inline Base.length(a::IdDistArray) = length(a.id)
+@inline Base.firstindex(a::IdDistArray) = 1
+@inline Base.lastindex(a::IdDistArray) = length(a)
+@inline Base.first(a::IdDistArray) = @inbounds a[begin]
+@inline Base.last(a::IdDistArray) = @inbounds a[end]
+@inline Base.eachindex(a::IdDistArray) = eachindex(a.id)
+@inline Base.popfirst!(a::IdDistArray) = popfirst!(a.id) => popfirst!(a.dist)
+@inline Base.pop!(a::IdDistArray) = pop!(a.id) => pop!(a.dist)
+@inline Base.getindex(a::IdDistArray, i::Integer) = @inbounds (a.id[i] => a.dist[i])
+
+@inline function Base.push!(a::IdDistArray, p::Pair)
+    push!(a.id, p.first)
+    push!(a.dist, p.second)
+    a
+end
+
+@inline function Base.setindex!(a::IdDistArray, p::Pair, i::Integer)
+    @inbounds a.id[i] = p.first
+    @inbounds a.dist[i] = p.second
+end
+
+@inline function Base.sizehint!(a::IdDistArray, sz::Integer)
+    sizehint!(a.id, sz)
+    sizehint!(a.dist, sz)
+    a
+end
+
+function reuse!(a::IdDistArray, k::Integer)
+    @assert k > 0
+    empty!(a.id)
+    empty!(a.dist)
+    sizehint!(a, k)
+    IdDistArray(a.id, a.dist, k)
 end
 
 """
@@ -17,15 +61,13 @@ size is reached. After this only the smallest items based on distance are preser
 """
 struct KnnResult{IdDistArray_} # <: AbstractVector{Tuple{IdType,DistType}}
     items::IdDistArray_
-    k::Int  # number of neighbors
 end
 
 function KnnResult(k::Integer)
     @assert k > 0
-    res = KnnResult(IdDistArray(Vector{Int32}(undef, 0), Vector{Float32}(undef, 0)), k)
-    sizehint!(res.items.id, k)
-    sizehint!(res.items.dist, k)
-    res
+    items = IdDistArray(Vector{Int32}(undef, 0), Vector{Float32}(undef, 0), k)
+    sizehint!(items, k)
+    KnnResult(items)
 end
 
 """
@@ -37,12 +79,11 @@ distribution of the items being inserted (it is expected just a few elements sma
 """
 function _shifted_fixorder!(res::KnnResult, sp::Int, ep::Int)
     ep == sp && return
-    id, dist = res.id, res.dist
+    id, dist = res.items.id, res.items.dist
     @inbounds i, d = id[ep], dist[ep]
     pos = _find_inspos(dist, sp, ep, d)
[…]
