### 2022-06-13 ebc441d
working KnnResultSet
