+++ src/knnresult.jl
@@ -1,7 +1,9 @@
 # This file is a part of SimilaritySearch.jl
-export KnnResult
-export maxlength, maxlength, getdist, getid, idview, distview, reuse!
+export KnnResult, KnnResultSet
+export maxlength, getdist, getid, idview, distview, reuse!
 
+###### KnnResult backends
+### array implementation, can grow efficiently
 struct IdDistArray
     id::Vector{Int32}
     dist::Vector{Float32}
@@ -53,6 +55,149 @@ function reuse!(a::IdDistArray, k::Integer)
 end
 
 """
+    _shifted_fixorder!(a::IdDistArray, sp, ep)
+
+Sorts the result in place; the possible element out of order is on the last entry always.
+It implements a kind of insertion sort that it is efficient due to the expected
+distribution of the items being inserted (it is expected just a few elements smaller than the current ones)
+"""
+function _shifted_fixorder!(a, sp::Int, ep::Int)
+    ep == sp && return
+    @inbounds i, d = a[ep]
+    pos = _find_inspos(a, sp, ep, d)
+    _shift_vector(a.id, pos, ep, i)
+    _shift_vector(a.dist, pos, ep, d)
+    nothing
+end
+
+@inline function _find_inspos(a, sp::Int, ep::Int, d::Float32)
+    while (mid = ep-sp) > 16
+        mid = sp + (mid >> 1)
+        d < getdist(a, mid) || break
+        ep = mid
+    end
+
+    while ep > sp
+        ep -= 1
+        d < getdist(a, ep) || return ep + 1
+    end
+
+    ep
+end
+
+@inline function _shift_vector(arr::Vector, sp::Int, ep::Int, val)
+    #=@inbounds while ep > sp;  arr[ep] = arr[ep-1]; ep -= 1; end=#
+    unsafe_copyto!(arr, sp+1, arr, sp, ep-sp)
+    @inbounds arr[sp] = val
+end
+
+############
+## Matrix based: uses preallocated matrices of identifiers and distances, can't grow. IdDistViews should not be stored (instead save the original KnnResultSet)
+
+struct KnnResultSet
+    id::Matrix{Int32}
+    dist::Matrix{Float32}
+    len::Vector{Int}
+end
+
+function KnnResultSet(k::Integer, m::Integer)
+    @assert k > 0 && m > 0
+
+    KnnResultSet(
+        Matrix{Int32}(undef, k, m),
+        Matrix{Float32}(undef, k, m),
+        zeros(Int, m)
+    )
+end
+
+function reuse!(s::KnnResultSet, i::Integer)
+    k = size(s.id, 1)
+    sp = (i - 1) * k + 1
+    KnnResult(IdDistViews(s, i, pointer(s.id, sp), pointer(s.dist, sp)))
+end
+
+"""
+    IdDistViews(ksearch::Integer)
+
+Creates a priority queue with fixed capacity (`ksearch`) representing a knn result set.
+It starts with zero items and grows with [`push!(res, id, dist)`](@ref) calls until `ksearch`
+size is reached. After this only the smallest items based on distance are preserved.
+"""
+struct IdDistViews
+    parent::KnnResultSet
+    i::Int
+    id::Ptr{Int32}
+    dist::Ptr{Float32}
+end
+
+@inline idview(a::IdDistViews) = @view a.parent.id[eachindex(a), a.i]
+@inline distview(a::IdDistViews) = @view a.parent.dist[eachindex(a), a.i]
+
+@inline getid(a::IdDistViews, i::Integer) = unsafe_load(a.id, i)
[…]
