### 2022-06-13 bdf4d7b
KnnResult for KnnResultSet and AbstractIdDist
