+++ benchmarks/knnresult.jl
@@ -1,6 +1,6 @@
 using SimilaritySearch, Random
 
-function main(res, n)
+function benchmark(res, n)
     # @info "===========", typeof(res), n
     push!(res, 1, rand())
 
@@ -23,12 +23,20 @@ function main(res, n)
     res
 end
 
-ksearch = 10
-n = 1000
-res = KnnResult(ksearch)
+function main(k, n)
+    res = KnnResult(k)
+    @info k n typeof(res)
+    @time benchmark(res, n)
 
-@timev main(res, n)
+    S = KnnResultSet(k, 1)
+    res = KnnResult(S, 1)
+    @info k n typeof(res)
+    @time benchmark(res, n)
+end
 
-n = 30_000_000
-res = KnnResult(ksearch)
-@timev main(res, n)
+@info "==== warming"
+main(10, 1000)
+
+@info "==== running benchmark"
+main(10, 30_000_000)
+
