### 2022-06-13 3d9d597
searchbatch without copying results
