### 2022-06-13 16f9cee
searchbatch without copying results
