+++ src/perf.jl
@@ -51,16 +51,3 @@ function macrorecall(goldlist::AbstractVector, reslist::AbstractVector)::Float64
 
     s / n
 end
-
-"""
-    timedsearchbatch(index, Q, ksearch::Integer; parallel=false)
-
-Computes the K nearest neigbors of each object in `Q` and returns two matrices, and the average search time in seconds.
-"""
-function timedsearchbatch(index, Q, ksearch::Integer; parallel=false, pools=getpools(index))
-    m = length(Q)
-    I = zeros(Int32, ksearch, m)
-    D = Matrix{Float32}(undef, ksearch, m)
-    t = @elapsed (I, D = searchbatch(index, Q, I, D; parallel, pools))
-    I, D, t/length(Q)
-end
