### 2022-06-13 2646427
bug fixes; Sat KnnResult pool is now a KnnResultSet; more tests
