+++ test/testresults.jl
@@ -46,7 +46,7 @@ function create_random_array(n, k)
     V, Vsorted
 end
 
-@testset "shifted vector-based result set" begin
+@testset "vector-based result set" begin
     k = 10
     res = KnnResult(k)
     V, Vsorted = create_random_array(50, k)
@@ -57,7 +57,7 @@ end
     testsorted(res, copy(Vsorted))
 end
 
-@testset "shifted vector-based result set" begin
+@testset "set of knn results" begin
     k = 10
     S = KnnResultSet(k, 1)
     res = KnnResult(S, 1)
@@ -69,3 +69,21 @@ end
     testsorted(res, copy(Vsorted))
 end
 
+@testset "equality" begin
+    k = 10
+    m = 1000
+    n = 10_000
+    S = KnnResultSet(k, m)
+    R = KnnResult(k)
+    for i in 1:m
+        test = rand(n)
+        V = KnnResult(S, i)
+        R = reuse!(R)
+        for (j, d) in enumerate(test)
+            push!(R, j, d)
+            push!(V, j, d)
+        end
+        
+        @test idview(V) == idview(R)
+    end
+end
