### 2022-06-13 358b64d
convenient functions for accessing KnnResultSet
