+++ src/searchgraph/visitedvertices.jl
@@ -1,26 +1,64 @@
 # This file is a part of SimilaritySearch.jl
 
-@inline visited(vstate::BitVector, i) = @inbounds vstate[i]
 
-@inline function visit!(vstate::BitVector, i)
+@inline function _bitindices(i::Integer)
+    i = UInt64(i) - one(UInt64)
+    (i >>> 6) + 1, (i & 63)
+end
+
+@inline function visited(vstate::Vector{UInt64}, i::Integer)
+    b, i = _bitindices(i)
+    @inbounds ((vstate[b] >>> i) & one(UInt64)) == one(UInt64)
+end
+
+@inline function visit!(vstate::Vector{UInt64}, i::Integer)
+    b, i = _bitindices(i)
+    @inbounds vstate[b] |= (one(UInt64) << i)
+end
+
+@inline function check_visited_and_visit!(vstate::Vector{UInt64}, i::Integer)
+    b, i = _bitindices(i)
+    v = ((vstate[b] >>> i) & one(UInt64)) == one(UInt64)
+    !v && ( @inbounds vstate[b] |= (one(UInt64) << i) )
+    v
+end
+
+#### VisitedVertices with BitVector
+
+@inline visited(vstate::BitVector, i::Integer) = @inbounds vstate[i]
+
+@inline function visit!(vstate::BitVector, i::Integer)
     @inbounds vstate[i] = true
 end
 
-@inline visited(vstate::Vector{UInt8}, i)::Bool = @inbounds vstate[i] == 1
+@inline function check_visited_and_visit!(vstate, i::Integer)
+    v = visited(vstate, i)
+    !v && visit!(vstate, i)
+    v
+end
 
-@inline function visit!(vstate::Vector{UInt8}, i, state)
+##### VisitedVertices with UInt8 vector
+
+@inline visited(vstate::Vector{UInt8}, i::Integer)::Bool = @inbounds (vstate[i] == 1)
+
+@inline function visit!(vstate::Vector{UInt8}, i::Integer)
     @inbounds vstate[i] = 1
 end
 
-@inline visited(vstate::Set{Int32}, i)::Bool = i ∈ vstate
+##### VisitedVertices with Set
 
-@inline function visit!(vstate::Set{Int32}, i)
+@inline visited(vstate::Set{Int32}, i::Integer)::Bool = i ∈ vstate
+
+@inline function visit!(vstate::Set{Int32}, i::Integer)
     @inbounds push!(vstate, i)
 end
+
+#### 
 
 const GlobalVisitedVertices = [BitArray(undef, 1)]  # initialized at __init__ function
 # const GlobalVisitedVertices = [Vector{UInt8}(undef, 1)]  # initialized at __init__ function
 # const GlobalVisitedVertices = [Set{Int32}()]  # initialized at __init__ function
+# const GlobalVisitedVertices = [Vector{UInt64}(undef, 32)]
 
 function __init__visitedvertices()
     for i in 2:Threads.nthreads()
@@ -33,14 +71,12 @@ function __init__visitedvertices()
 end
 
 @inline function _init_vv(v::AbstractVector, n)
-    # length(v) < n &&
     resize!(v, n)
     fill!(v, 0)
     v
 end
 
 @inline function _init_vv(v::BitVector, n)
-    # length(v) < n &&
     resize!(v, n)
     fill!(v.chunks, 0)
     v
@@ -48,5 +84,12 @@ end
 
 @inline function _init_vv(v::Set, n)
     empty!(v)
+    v
+end
+
+@inline function _init_vv(v::Vector{UInt64}, n)
+    n64 = 1 + ((n-1) >>> 6)
+    n64 != length(v) && resize!(v, n64)
+    fill!(v, 0)
     v
 end
