+++ src/searchgraph/beamsearch.jl
@@ -31,8 +31,7 @@ end
 ### local search algorithm
 
 @inline function beamsearch_queue(index::SearchGraph, q, res::KnnResult, objID, vstate)
-    visited(vstate, objID) && return 0
-    visit!(vstate, objID)
+    check_visited_and_visit!(vstate, objID) && return 0
     @inbounds push!(res, objID, evaluate(index.dist, q, index[objID]))
     1
 end
@@ -66,8 +65,7 @@ function beamsearch_inner(bs::BeamSearch, index::SearchGraph, q, res::KnnResult,
         prev_dist > maximum(res) && break
         sp += 1
         for childID in index.links[prev_id]
-            visited(vstate, childID) && continue
-            visit!(vstate, childID)
+            check_visited_and_visit!(vstate, childID) && continue
             d = convert(Float32, evaluate(index.dist, q, index[childID]))
             p = childID => d
             push!(res, p)
