### 2022-06-14 98b55fc
CompatHelper: add new compat entry for PProf at version 2, (keep existing compat)
