### 2022-06-14 c9cbcdd
in-the-house BitVector replacement to factorize operations; simplification of the allknn method; the two versions of knnresults are now fully working
