+++ src/allknn.jl
@@ -4,78 +4,39 @@ export allknn
 
 """
     allknn(g::AbstractSearchContext, k::Integer; minbatch=0, pools=getpools(g)) -> knns, dists
+    allknn(g, knns::Matrix{Int32}, dists::Matrix{Float32}; minbatch=0, pools=getpools(g)) -> knns, dists
+    allknn(g, R::KnnResultSet; minbatch=0, pools=getpools(g)) -> knns, dists
 
 Computes all the k nearest neighbors (all vs all) using the index `g`. It removes self references.
 
-Parameters:
+# Parameters:
 
 - `g`: the index
-- `k`: the number of neighbors to retrieve
+- Query specification and result:
+    - `k`: the number of neighbors to retrieve
+    - `knns`: an uninitialized integer matrix of (k, n) size for storing the `k` nearest neighbors of the `n` elements
+    - `dists`: an uninitialized floating point matrix of (k, n) size for storing the `k` nearest distances of the `n` elements
+    - `R`: an uninitialized `KnnResultSet` (contains `knns` and `dists` internally, along with the lenghts of the results)
 - `minbatch`: controls how multithreading is used for evaluating configurations, see [`getminbatch`](@ref)
 - `pools`: A pools object, dependent of `g`
 
-Returns:
+# Returns:
 
 - `knns` a (k, n) matrix of identifiers; the i-th column corresponds to the i-th object in the dataset.
     Zeros can happen to the end of each column meaning that the retrieval was less than the desired `k`
 - `dists` a (k, n) matrix of distances; the i-th column corresponds to the i-th object in the dataset.
     Zero values in `knns` should be ignored in `dists`
 
+These are the same buffers passed as arguments in some of the function methods.
+
+# Note:
+This function was introduced in `v0.8` series, and removes self references automatically.
+In `v0.9` the self reference is kept since removing from the algorithm introduces unreasonable overhead for some case of use; however,
+it ensures that the self-reference is in the first entry of the result set.
 """
-function allknn(g::AbstractSearchContext, k::Integer; minbatch=0, pools=getpools(g))
-    n = length(g)
-    R = KnnResultSet(k, n)
-    allknn(g, R; minbatch, pools)
-end
+allknn(g::AbstractSearchContext, k::Integer; minbatch=0, pools=getpools(g)) = allknn(g, KnnResultSet(k, length(g)); minbatch, pools)
+allknn(g::AbstractSearchContext, knns::Matrix{Int32}, dists::Matrix{Float32}; minbatch=0, pools=getpools(g)) = allknn(g, KnnResultSet(knns, dists); minbatch, pools)
 
-function _allknn_loop(g::SearchGraph, i::Integer, res::KnnResult, pools)
-    vstate = getvstate(length(g), pools)
-    # visit!(vstate, i)
-    @inbounds _allknn_loop_barrier(g, i, g[i], res, g.links[i], pools, vstate)
-end
-
-function _allknn_loop_barrier(g::SearchGraph, i, c, res, hints, pools, vstate)
-    k = maxlength(res)
-
-    # the loop helps to overcome when the current nn is in a small clique (smaller the the desired k)
-    for h in hints # hints
-        visited(vstate, h) && continue
-        search(g.search_algo, g, c, res, h, pools; vstate)
-        length(res) == k && break
-    end
-    # search(g.search_algo, g, c, res, hints, pools; vstate)
-
-    #=if length(res) < k
-        for _ in 1:k
-            h = rand(1:length(g))
-            visited(vstate, h) && continue
-            search(g.search_algo, g, c, res, h, pools; vstate)
-            length(res) == k && break
-        end
-    end=#
-end
-
[…]
