### 2022-06-14 eeb1c54
changing muladd istead of fma improves a bit the speed of evaluate
