+++ src/searchgraph/beamsearch.jl
@@ -61,8 +61,9 @@ function beamsearch_inner(bs::BeamSearch, index::SearchGraph, q, res::KnnResult,
     sp = 1
 
     @inbounds while sp <= length(beam)
-        prev_id, prev_dist = beam[sp]
-        prev_dist > maximum(res) && break
+        # prev_id, prev_dist = beam[sp]
+        # prev_dist > maximum(res) && break
+        prev_id = getid(beam, sp)
         sp += 1
         for childID in index.links[prev_id]
             check_visited_and_visit!(vstate, childID) && continue
