### 2022-06-15 7af9aa4
removes fixing self reference to the first element (should be checked by applications); set restarts to hints
