### 2022-06-15 90c9648
merge with main (without generic KnnResult)
