+++ src/knnresult.jl
@@ -1,6 +1,6 @@
 # This file is a part of SimilaritySearch.jl
 export KnnResult
-export maxlength, maxlength, getdist, getid, idview, distview, reuse!
+export maxlength, getdist, getid, idview, distview, reuse!
 
 """
     KnnResult(ksearch::Integer)
@@ -124,7 +124,7 @@ Returns a result set and a new initial state; reuse the memory buffers
 @inline function reuse!(res::KnnResult, k::Integer=res.k)
     @assert k > 0
     empty!(res.id); empty!(res.dist)
-    sizehint!(res.id, k), sizehint!(res.dist, k)
+    sizehint!(res.id, k); sizehint!(res.dist, k)
     KnnResult(res.id, res.dist, k)
 end
 
