### 2022-06-15 b2282aa
merge with copying-knnresult-again branch
