+++ src/knnresult.jl
@@ -2,30 +2,8 @@
 export KnnResult
 export maxlength, getdist, getid, idview, distview, reuse!
 
-@inline idview(a::IdDistArray) = a.id
-@inline distview(a::IdDistArray) = a.dist
-@inline getid(a::IdDistArray, i::Integer) = @inbounds a.id[i]
-@inline getdist(a::IdDistArray, i::Integer) = @inbounds a.dist[i]
-@inline maxlength(a::IdDistArray) = a.k
-@inline Base.maximum(a::IdDistArray) = last(a.dist)
-@inline Base.argmax(a::IdDistArray) = last(a.id)
-@inline Base.minimum(a::IdDistArray) = first(a.dist)
-@inline Base.argmin(a::IdDistArray) = first(a.id)
-@inline Base.length(a::IdDistArray) = length(a.id)
-@inline Base.firstindex(a::IdDistArray) = 1
-@inline Base.lastindex(a::IdDistArray) = length(a)
-@inline Base.first(a::IdDistArray) = @inbounds a[begin]
-@inline Base.last(a::IdDistArray) = @inbounds a[end]
-@inline Base.eachindex(a::IdDistArray) = eachindex(a.id)
-@inline Base.popfirst!(a::IdDistArray) = popfirst!(a.id) => popfirst!(a.dist)
-@inline Base.pop!(a::IdDistArray) = pop!(a.id) => pop!(a.dist)
-@inline Base.getindex(a::IdDistArray, i::Integer) = @inbounds (a.id[i] => a.dist[i])
-
-@inline function Base.push!(a::IdDistArray, p::Pair)
-    push!(a.id, p.first)
-    push!(a.dist, p.second)
-    a
-end
+"""
+    KnnResult(ksearch::Integer)
 
 Creates a priority queue with fixed capacity (`ksearch`) representing a knn result set.
 It starts with zero items and grows with [`push!(res, id, dist)`](@ref) calls until `ksearch`
@@ -37,13 +15,7 @@ struct KnnResult # <: AbstractVector{Tuple{IdType,DistType}}
     k::Int  # number of neighbors
 end
 
-@inline function Base.sizehint!(a::IdDistArray, sz::Integer)
-    sizehint!(a.id, sz)
-    sizehint!(a.dist, sz)
-    a
-end
-
-function reuse!(a::IdDistArray, k::Integer)
+function KnnResult(k::Integer)
     @assert k > 0
     res = KnnResult(Vector{Int32}(undef, 0), Vector{Float32}(undef, 0), k)
     sizehint!(res.id, k)
@@ -52,31 +24,33 @@ function reuse!(a::IdDistArray, k::Integer)
 end
 
 """
-    _shifted_fixorder!(a::IdDistArray, sp, ep)
+    _shifted_fixorder!(res::KnnResult, sp, ep)
 
 Sorts the result in place; the possible element out of order is on the last entry always.
 It implements a kind of insertion sort that it is efficient due to the expected
 distribution of the items being inserted (it is expected just a few elements smaller than the current ones)
 """
-function _shifted_fixorder!(a, sp::Int, ep::Int)
+function _shifted_fixorder!(res::KnnResult, sp::Int, ep::Int)
     ep == sp && return
-    @inbounds i, d = a[ep]
-    pos = _find_inspos(a, sp, ep, d)
-    _shift_vector(a.id, pos, ep, i)
-    _shift_vector(a.dist, pos, ep, d)
+    id, dist = res.id, res.dist
+    @inbounds i, d = id[ep], dist[ep]
+    pos = _find_inspos(dist, sp, ep, d)
+    _shift_vector(id, pos, ep, i)
+    _shift_vector(dist, pos, ep, d)
+
     nothing
 end
 
[…]
