+++ benchmarks/allknn-lv.jl
@@ -1,28 +1,35 @@
 
-using SimilaritySearch, LoopVectorization, StrideArrays, LinearAlgebra
+using SimilaritySearch, LoopVectorization, StrideArrays, LinearAlgebra, Random
 
 struct NormalizedCosineDistanceLV{Dim} <: SemiMetric
 end
 
 function SimilaritySearch.evaluate(::NormalizedCosineDistanceLV{Dim}, u::AbstractVector{T}, v::AbstractVector{T}) where {Dim,T}
     d = zero(T)
-    @turbo inline=true unroll=2 thread=1 for i in 1:Dim
+    @turbo unroll=2 thread=1 for i in 1:Dim
         d = muladd(u[i], v[i], d)
+        # d += u[i] * v[i]
     end
 
     one(T) - d
 end
 
-function create_database()
-    X = rand(Float32, 8, 100_000)
+function create_database(dim=100, n=100_000, filled=8)
+    X = zeros(Float32, dim, n)
+    for i in 1:n
+        rand!(view(X, 1:filled, i))
+    end
+    
     for c in eachcol(X) normalize!(c) end
-    MatrixDatabase(StrideArray(X, StaticInt.(size(X))))
+    # MatrixDatabase(X), NormalizedCosineDistanceLV{dim}()
+    MatrixDatabase(StrideArray(X, StaticInt.(size(X)))), NormalizedCosineDistanceLV{dim}()   # slow compilation, fast computation
+    # MatrixDatabase(StrideArray(X, size(X))), NormalizedCosineDistanceLV{8}()
+    #MatrixDatabase(X), NormalizedCosineDistance()   # fast compilation, a bit slower
 end
 
 function main()
     @info "this benchmark is intended to work with multithreading enabled julia sessions"
-    db = create_database()
-    dist = NormalizedCosineDistanceLV{8}()
+    db, dist = create_database()
     k = 32
     @info "----- computing gold standard"
     goldsearchtime = @elapsed gI, gD = allknn(ExhaustiveSearch(; db, dist), k)
