### 2022-06-16 796dc9e
adds v0.9 notes in README.md; increases version
