+++ Project.toml
@@ -8,7 +8,6 @@ Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 Intersections = "6d0fbc77-481e-44fe-9ace-1653ddcdea34"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-PProf = "e4faabce-9ead-11e9-39d9-4379958e3056"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
