### 2022-06-17 e58a8c9
moves parallel_block's defaults computation to a function; UInt64 visited vertices now only fills with zeros the necessary elements
