+++ src/searchgraph/insertions.jl
@@ -1,11 +1,20 @@
 # This file is a part of SimilaritySearch.jl
 
 """
+    get_parallel_block()
+
+Used by SearchGraph insertion functions to solve `find_neighborhood` in blocks. Small blocks are better to ensure quality; faster constructions will be achieved if `parallel_block` is a multiply of `Threads.nthreads()`
+
+"""
+get_parallel_block() = Threads.nthreads() == 1 ? 1 : 8 * Threads.nthreads()
+
+
+"""
     append!(
         index::SearchGraph,
         db;
         neighborhood=Neighborhood(),
-        parallel_block=(Threads.nthreads() == 1 ? 1 : 8 * Threads.nthreads()),
+        parallel_block=get_parallel_block(),
         parallel_minimum_first_block=parallel_block,
         callbacks=SearchGraphCallbacks(),
         pools=getpools(index)
@@ -36,7 +45,7 @@ function Base.append!(
         index::SearchGraph,
         db;
         neighborhood=Neighborhood(),
-        parallel_block=(Threads.nthreads() == 1 ? 1 : 8 * Threads.nthreads()),
+        parallel_block=get_parallel_block(),
         parallel_minimum_first_block=parallel_block,
         callbacks=SearchGraphCallbacks(),
         pools=getpools(index)
@@ -63,7 +72,7 @@ function Base.append!(
 end
 
 """
-    index!(index::SearchGraph; parallel_block=(Threads.nthreads() == 1 ? 1 : 8 Threads.nthreads(), parallel_minimum_first_block=parallel_block, callbacks=Sear* chGraphCallbacks()))
+    index!(index::SearchGraph; parallel_block=get_parallel_block(), parallel_minimum_first_block=parallel_block, callbacks=SearchGraphCallbacks())
 
 Indexes the already initialized database (e.g., given in the constructor method). It can be made in parallel or sequentially.
 The arguments are the same than `append!` function but using the internal `index.db` as input.
@@ -72,7 +81,7 @@ The arguments are the same than `append!` function but using the internal `index
 function index!(
         index::SearchGraph;
         neighborhood=Neighborhood(),
-        parallel_block=(Threads.nthreads() == 1 ? 1 : 8 * Threads.nthreads()),
+        parallel_block=get_parallel_block(),
         parallel_minimum_first_block=parallel_block,
         callbacks=SearchGraphCallbacks(),
         pools=getpools(index)
@@ -108,8 +117,7 @@ function _sequential_append_loop!(index::SearchGraph, neighborhood::Neighborhood
 end
 
 function _connect_links(index, sp, ep)
-    minbatch = getminbatch(0, ep-sp+1)
-    @batch minbatch=minbatch per=thread for i in sp:ep
+    @batch minbatch=getminbatch(0, ep-sp+1) per=thread for i in sp:ep
         @inbounds for id in index.links[i]
             lock(index.locks[id])
             try
@@ -122,14 +130,15 @@ function _connect_links(index, sp, ep)
     end
 end
 
-function _parallel_append_loop!(index::SearchGraph, neighborhood::Neighborhood, pools::SearchGraphPools, sp, n, parallel_block, callbacks)
+function _parallel_append_loop!(index::SearchGraph, neighborhood::Neighborhood, pools::SearchGraphPools, sp, n, parallel_block::Integer, callbacks)
     while sp < n
         ep = min(n, sp + parallel_block)
         index.verbose && rand() < 0.01 && println(stderr, "appending chunk ", (sp=sp, ep=ep, n=n), " ", Dates.now())
 
         # searching neighbors
         # @show length(index.links), length(index.db), length(db), length(index.locks), length(index), sp, ep
-        @batch minbatch=getminbatch(0, ep-sp+1) per=thread for i in sp:ep
[…]
