### 2022-06-17 e0c8804
simplified getminbatch; default policy on allknn is to hint on a random neighbor (simplification)
