### 2022-06-17 499ef6b
simplify SearchGraphCallbacks; also now accepts nothing to skip inline hyperparameters; accepts verbose=false easily
