### 2022-06-17 ef1324a
renames AbstractSearchContext by AbstractSearchIndex; the name was kept in v0.8 but since the v0.9 series is starting, the old type name will be treated as a bug in favor of the more precise type name (semantically)
