+++ src/searchgraph/graph.jl
@@ -25,7 +25,7 @@ abstract type Callback end
 ### Basic operations on the index
 
 """
-    struct SearchGraph <: AbstractSearchContext
+    struct SearchGraph <: AbstractSearchIndex
 
 SearchGraph index. It stores a set of points that can be compared through a distance function `dist`.
 The performance is determined by the search algorithm `search_algo` and the neighborhood policy.
@@ -36,7 +36,7 @@ It supports callbacks to adjust parameters as insertions are made.
 Note: Parallel insertions should be made through `append!` or `index!` function with `parallel_block > 1`
 
 """
-@with_kw struct SearchGraph{DistType<:SemiMetric, DataType<:AbstractDatabase, SType<:LocalSearchAlgorithm}<:AbstractSearchContext
+@with_kw struct SearchGraph{DistType<:SemiMetric, DataType<:AbstractDatabase, SType<:LocalSearchAlgorithm}<:AbstractSearchIndex
     dist::DistType = SqL2Distance()
     db::DataType = VectorDatabase()
     links::Vector{Vector{Int32}} = Vector{Int32}[]
