### 2022-06-19 1c3461a
getminbatch resolves -1 as n
