### 2022-06-19 75c1216
add JULIA_NUM_THREADS in ci.yml
