### 2022-06-19 baaa0a2
add JULIA_NUM_THREADS in ci.yml
