### 2022-06-19 9342645
adds assertions on perf.jl
