### 2022-06-19 122437f
adds assertions on perf.jl
