### 2022-06-19 219aae7
adds more tests for allknn
