+++ test/testallknn.jl
@@ -5,18 +5,21 @@ using Test, SimilaritySearch
 @testset "allknn" begin
     k = 5
     dist = L2Distance()
-    X = MatrixDatabase(rand(Float32, 4, 100))
+    n = 100
+    X = MatrixDatabase(rand(Float32, 4, n))
     G = SearchGraph(; db=X, dist)
     index!(G)
     optimize!(G, MinRecall(0.95))
     gI, gD = allknn(G, k)
 
+    @test size(gI) == size(gD) == (k, n)
     E = ExhaustiveSearch(; db=X, dist)
     eI, eD = allknn(E, k)
+    @test size(eI) == size(eD) == (k, n)
 
     P = ParallelExhaustiveSearch(; db=X, dist)
     pI, pD = allknn(P, k)
-
+    @test size(pI) == size(pD) == (k, n)
     @test macrorecall(eI, gI) > 0.8
     @test macrorecall(eI, pI) > 0.99
 end
