### 2022-06-19 a937fe0
adds more tests for allknn
