### 2022-06-19 1056af2
adds more tests
