+++ test/testsearchgraph.jl
@@ -27,7 +27,7 @@ using Test
         append!(graph, db; neighborhood, parallel_block=8, callbacks=SearchGraphCallbacks(nothing))
         @test n == length(db) == length(graph)
         searchtime = @elapsed I, D = searchbatch(graph, queries, ksearch)
-        @test size(I) == size(D) == (ksearch, m)
+        @test size(I) == size(D) == (ksearch, m) == size(goldI)
         #@info sort!(length.(graph.links), rev=true)
         @show goldD[:, 1]
         @show D[:, 1]
@@ -48,7 +48,7 @@ using Test
     @info "---- starting ParetoRadius optimization ---"
     optimize!(graph, ParetoRadius())
     searchtime = @elapsed I, D = searchbatch(graph, queries, ksearch)
-    @test size(I) == size(D) == (ksearch, m)
+    @test size(I) == size(D) == (ksearch, m) == size(goldI)
     recall = macrorecall(goldI, I)
     @info "ParetoRadius:> queries per second: ", m/searchtime, ", recall:", recall
     @info graph.search_algo
@@ -58,6 +58,7 @@ using Test
     @info "---- starting ParetoRecall optimization ---"
     optimize!(graph, ParetoRecall())
     searchtime = @elapsed I, D = searchbatch(graph, queries, ksearch)
+    @test size(I) == size(D) == (ksearch, m) == size(goldI)
     recall = macrorecall(goldI, I)
     @info "ParetoRecall:> queries per second: ", m/searchtime, ", recall:", recall
     @info graph.search_algo
@@ -66,8 +67,10 @@ using Test
     @info "========================= REBUILD TEST ======================"
     graph = SearchGraph(; db, dist, verbose=false)
     index!(graph; callbacks=SearchGraphCallbacks(MinRecall(0.9)))
+    @test n == length(db) == length(graph)
     optimize!(graph, MinRecall(0.9); queries)
     searchtime = @elapsed I, D = searchbatch(graph, queries, ksearch)
+    @test size(I) == size(D) == (ksearch, m) == size(goldI)
     recall = macrorecall(goldI, I)
     @info "testing without additional optimizations: queries per second:", m/searchtime, ", recall: ", recall
     @info graph.search_algo
@@ -75,8 +78,10 @@ using Test
 
     @info "========================= rebuild process =========================="
     graph = rebuild(graph)
+    @test n == length(db) == length(graph)
     optimize!(graph, MinRecall(0.9); queries)  # using the actual dataset makes prone to overfitting hyperparameters (more noticeable in rebuilt indexes)
     searchtime_ = @elapsed I, D = searchbatch(graph, queries, ksearch)
+    @test size(I) == size(D) == (ksearch, m) == size(goldI)
     recall_ = macrorecall(goldI, I)
 
     @info "-- old vs rebuild> searchtime: $searchtime vs $searchtime_; recall: $recall vs $recall_"
@@ -89,6 +94,7 @@ using Test
     goldI, goldD = searchbatch(seq, queries, ksearch)
     graph = SearchGraph(; db, dist, verbose=false)
     buildtime = @elapsed index!(graph)
+    @test n == length(db) == length(graph)
     searchtime = @elapsed I, D = searchbatch(graph, queries, ksearch)
     recall = macrorecall(goldI, I)
     @info "buildtime", buildtime
