+++ test/testallknn.jl
@@ -7,7 +7,7 @@ using Test, SimilaritySearch
     dist = L2Distance()
     n = 100
     X = MatrixDatabase(rand(Float32, 4, n))
-    G = SearchGraph(; db=X, dist)
+    G = SearchGraph(; db=X, dist, verbose=false)
     index!(G)
     @test length(G) == n
     optimize!(G, MinRecall(0.95))
