### 2022-06-19 dd1536e
removes threading in github action test
