+++ .github/workflows/ci.yml
@@ -13,8 +13,6 @@ jobs:
   test:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
     runs-on: ${{ matrix.os }}
-    env:
-      JULIA_NUM_THREADS: auto
     strategy:
       fail-fast: false
       matrix:
