### 2022-06-20 86814fe
visited vertices as a struct
