+++ src/closestpair.jl
@@ -31,7 +31,7 @@ end
 function search_hint(G::SearchGraph, i::Integer, pools)
     res = getknnresult(3, pools)
     vstate = getvstate(length(G), pools)
-    visit!(vstate, i)
+    visit!(vstate, convert(UInt64, i))
     search(G.search_algo, G, G[i], res, rand(G.links[i]), pools; vstate)
     argmin(res), minimum(res)
 end
