### 2022-06-22 3601380
go back to previous strategy in allknn
