### 2022-06-23 eb600c7
small speed ups
