### 2022-06-23 b0b4209
db refactoring; change allknn-lv.jl to compare with allknn.jl
