+++ benchmarks/allknn-lv.jl
@@ -1,3 +1,4 @@
+# This file is a part of SimilaritySearch.jl
 
 using SimilaritySearch, LoopVectorization, StrideArrays, LinearAlgebra, Random
 
@@ -14,14 +15,14 @@ function SimilaritySearch.evaluate(::NormalizedCosineDistanceLV{Dim}, u::Abstrac
     one(T) - d
 end
 
-function create_database(dim=100, n=100_000, filled=8)
+function create_database(dim=100, filled=8, n=100_000)
     X = zeros(Float32, dim, n)
     for i in 1:n
         rand!(view(X, 1:filled, i))
     end
     
     for c in eachcol(X) normalize!(c) end
-    # MatrixDatabase(X), NormalizedCosineDistanceLV{dim}()
+    #MatrixDatabase(X), NormalizedCosineDistanceLV{dim}()
     MatrixDatabase(StrideArray(X, StaticInt.(size(X)))), NormalizedCosineDistanceLV{dim}()   # slow compilation, fast computation
     # MatrixDatabase(StrideArray(X, size(X))), NormalizedCosineDistanceLV{8}()
     #MatrixDatabase(X), NormalizedCosineDistance()   # fast compilation, a bit slower
@@ -29,16 +30,20 @@ end
 
 function main()
     @info "this benchmark is intended to work with multithreading enabled julia sessions"
-    db, dist = create_database()
+    db, dist = create_database(8, 8)
     k = 32
     @info "----- computing gold standard"
+    GC.enable(false)
     goldsearchtime = @elapsed gI, gD = allknn(ExhaustiveSearch(; db, dist), k)
+    GC.enable(true)
     @info "----- computing search graph"
     H = SearchGraph(; db, dist, verbose=false)
     index!(H; parallel_block=256)
     optimize!(H, MinRecall(0.9), ksearch=k)
     # prune!(RandomPruning(12), H)
+    GC.enable(false)
     searchtime = @elapsed hI, hD = allknn(H, k; minbatch=32)
+    GC.enable(true)
     n = length(db)
     @info "gold:" (; n, goldsearchtime, qps=n/goldsearchtime)
     @info "searchgraph:" (; n, searchtime, qps=n / searchtime)
