+++ src/db/matrixdatabase.jl
@@ -16,7 +16,7 @@ struct MatrixDatabase{M<:AbstractMatrix} <: AbstractDatabase
     matrix::M  # abstract matrix
 end
 
-@inline Base.eltype(db::MatrixDatabase) = AbstractVector{eltype(db.matrix)}
+@inline Base.eltype(db::MatrixDatabase) = AbstractVector{eltype(db.matrix)} 
 
 """
     MatrixDatabase(V::MatrixDatabase)    
@@ -25,6 +25,8 @@ Creates another `MatrixDatabase` from another `MatrixDatabase`. They will share 
 """
 MatrixDatabase(V::MatrixDatabase) = MatrixDatabase(V.matrix)
 
+@inline Base.getindex(db::MatrixDatabase{<:StrideArray}, i::Integer) = view(db.matrix, :, i)
+@inline Base.getindex(db::MatrixDatabase{<:DenseArray}, i::Integer) = PtrArray(view(db.matrix, :, i))
 @inline Base.getindex(db::MatrixDatabase, i::Integer) = view(db.matrix, :, i)
 @inline Base.setindex!(db::MatrixDatabase, value, i::Integer) = @inbounds (db.matrix[:, i] .= value)
 @inline Base.push!(db::MatrixDatabase, v) = error("push! is not supported for MatrixDatabase, please see DynamicMatrixDatabase")
