### 2022-06-23 66197ae
replaces array copying operations by PtrArray arr2 .= arr1 (it goes faster)
