### 2022-06-23 710c49e
adds StrideArraysCore to [compat]
