+++ Project.toml
@@ -22,7 +22,8 @@ Parameters = "0.12"
 Polyester = "0.6"
 SearchModels = "0.3"
 StatsBase = "0.30.0, 0.32.0, 0.33"
-julia = "1.6"
+StrideArraysCore = "0.3"
+julia = "1.6, 1.7"
 
 [extras]
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
