### 2022-06-25 61e86a0
use PtrArray in some convenient ways to speedup ops
