### 2022-06-25 2faaa48
adds Aqua tests
