+++ src/searchgraph/SearchGraph.jl
@@ -1,9 +1,8 @@
 # This file is a part of SimilaritySearch.jl
 
-export LocalSearchAlgorithm, SearchGraph, SearchGraphPools, SearchGraphCallbacks, VisitedVertices, index!, push_item!
+export LocalSearchAlgorithm, SearchGraph, SearchGraphPools, SearchGraphCallbacks, index!, push_item!
 export Neighborhood, IdentityNeighborhood, DistalSatNeighborhood, SatNeighborhood
 export find_neighborhood, push_neighborhood!
-export callbacks
 export BeamSearch, BeamSearchSpace, Callback
 export KDisjointHints, DisjointHints, RandomHints
 export RandomPruning, KeepNearestPruning, SatPruning, prune!
