+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.9.3"
 [deps]
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
-Intersections = "6d0fbc77-481e-44fe-9ace-1653ddcdea34"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Polyester = "f517fe37-dbe3-4b94-8317-1923a5111588"
@@ -16,8 +15,8 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StrideArraysCore = "7792a7ef-975c-4747-a70f-980b88e8d1da"
 
 [compat]
+Aqua = "0.5"
 Distances = "0.10"
-Intersections = "0.2.8"
 Parameters = "0.12"
 Polyester = "0.6"
 SearchModels = "0.3"
@@ -26,8 +25,9 @@ StrideArraysCore = "0.3"
 julia = "1.6, 1.7"
 
 [extras]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "SparseArrays"]
+test = ["Test", "SparseArrays", "Aqua"]
