### 2022-06-27 #4442 @ahwillia
Ok, I do have one very small suggestion for @sadit -- it would be nice to add some links at the top of the GitHub README to other alternative libraries in Julia like [`NearestNeighbors.jl`](https://github.com/KristofferC/NearestNeighbors.jl) which seems to also be actively maintained. My understanding is that `SimilaritySearch.jl` is better suited to find approximate solutions on very large datasets than existing options. Many users will not read the pdf compiled for JOSS so it would be nice to briefly summarize these alternative libraries on the main page and how `SimilaritySearch.jl` is different.
