### 2022-06-27 #4442 @kose-y
@sadit -- It looks like a great package, and the authors did a thorough job with the submission. However, the example code provided in the paper has a minor error:

Code line 20:
```julia
example()
```
must have a value of $k$ as an argument.

@jbytecode I have a question about the package version consistency. The submitted version of the package is `v0.8.20`, while the installation instruction would install the newest version, `v0.9.3`. There has been a breaking change between them, so the example code in the paper would not work correctly with the newer version. Specifically, the keyword argument `parallel` has been removed in `v0.9`. Should the authors provide the installation instruction for `v0.8.20` for their installation instruction in the paper?
