### 2022-06-28 #4442 @kose-y
Also, the citation block ending on Line 21 has an extra semicolon.
