### 2022-06-28 f373aef
add Core at test ambiguities
