+++ joss/example/Project.toml
@@ -6,9 +6,13 @@ LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 MLDatasets = "eb30cadb-4394-5ae3-aed4-317e484a6458"
 SimilaritySearch = "053f045d-5466-53fd-b400-a066f88fe02a"
 StrideArrays = "d1fa6d79-ef01-42a6-86c9-f7c551f8593b"
+StrideArraysCore = "7792a7ef-975c-4747-a70f-980b88e8d1da"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
 [compat]
-MLDatasets = "0.6.0"
-SimilaritySearch = "0.8.20"
 JSON = "0.21"
+MLDatasets = "0.6.0"
+SimilaritySearch = "0.9.3"
+LoopVectorization = "0.12"
+StrideArrays = "0.1"
+StrideArraysCore = "0.3"
