### 2022-06-28 324174f
add notes about Similarity Search alternatives in README; fix typo in paper.md HNSW
