+++ joss/paper.md
@@ -46,7 +46,7 @@ Our `SearchGraph` is based on the Navigable Small World (NSW) graph index [@malk
 @malkov2014approximate add a hierarchical structure to the NSW to create the Hierarchical NSW (HNSW) search structure. This index is a central component of the [`hnswlib`](https://github.com/nmslib/hnswlib) and the [`nmslib`](https://github.com/nmslib/nmslib) libraries. Along with the HNSW, the [`faiss`](https://github.com/facebookresearch/faiss) library also provides a broad set of efficient implementations of metric, hashing, and product quantization indexes. @nndescent11 introduces the NN Descent method, which uses the graph of neighbors as index structure; it is the machinery behind [`PyNNDescent`](https://github.com/lmcinnes/pynndescent), which is behind the fast computation of UMAP non-linear low dimensional projection.^[<https://github.com/lmcinnes/umap>.]
 @scann2020 introduced the _SCANN_ index for inner product-based metrics and Euclidean distance, available at the [SCANN repository](https://github.com/google-research/google-research/tree/master/scann) based on hashing.
 
-Currently, there exists some packages dedicated to nearest neighbor search, for instance we have [`NearestNeighbors.jl`](https://github.com/KristofferC/NearestNeighbors.jl), [`Rayuela.jl`](https://github.com/una-dinosauria/Rayuela.jl), [`HSNW.jl`](https://github.com/JuliaNeighbors/HNSW.jl), and a wrapper for the FAISS library, [`Faiss.jl`](https://github.com/zsz00/Faiss.jl), among other efforts.
+Currently, there exists some packages dedicated to nearest neighbor search, for instance we have [`NearestNeighbors.jl`](https://github.com/KristofferC/NearestNeighbors.jl), [`Rayuela.jl`](https://github.com/una-dinosauria/Rayuela.jl), [`HNSW.jl`](https://github.com/JuliaNeighbors/HNSW.jl), and a wrapper for the FAISS library, [`Faiss.jl`](https://github.com/zsz00/Faiss.jl), among other efforts.
 
 
 # Main features of `SimilaritySearch`
