### 2022-06-28 b3db504
adds alternatives in README.md
