+++ README.md
@@ -6,7 +6,7 @@
 # SimilaritySearch.jl
 
 
-SimilaritySearch.jl is a library for nearest neighbor search. In particular, it contains the implementation for `SearchGraph`, a fast and flexible search index using any metric function. It is designed to support multithreading in most of its main functions and structures.
+SimilaritySearch.jl is a library for nearest neighbor search. In particular, it contains the implementation for `SearchGraph,` a fast and flexible search index using any metric function. It is designed to support multithreading in most of its functions and structures.
 
 The package provides the following indexes:
 
@@ -19,10 +19,18 @@ The main set of functions are:
 - `search`: Solves a single query.
 - `searchbatch`: Solves a set of queries.
 - `allknn`: Computes the $k$ nearest neighbors for all elements in an index.
-- `neardup`: Removes a near duplicates from a metric dataset.
+- `neardup`: Removes near-duplicates from a metric dataset.
 - `closestpair`: Computes the closest pair in a metric dataset.
 
-The precise definitions of these functions and the full set of functions and structures can be found in the [documentation](https://sadit.github.io/SimilaritySearch.jl/dev).
+The precise definitions of these functions and the complete set of functions and structures can be found in the [documentation](https://sadit.github.io/SimilaritySearch.jl/dev).
+
+# Similarity search _ecosystem_ in Julia
+Currently, there exists several packages dedicated to nearest neighbor search, for instance we have [`NearestNeighbors.jl`](https://github.com/KristofferC/NearestNeighbors.jl), [`RegionTrees.jl`](https://github.com/rdeits/RegionTrees.jl), and [`JuliaNeighbors`](https://github.com/JuliaNeighbors) implement search structures like [kd-trees](https://en.wikipedia.org/wiki/K-d_tree), [ball trees](https://en.wikipedia.org/wiki/Ball_tree), [quadtrees](https://en.wikipedia.org/wiki/Quadtree), [octrees](https://en.wikipedia.org/wiki/Octree), [bk-trees](https://en.wikipedia.org/wiki/BK-tree), [vp-tree](https://en.wikipedia.org/wiki/Vantage-point_tree) and other multidimensional and metric structures. These structures work quite well for low dimensional data since they are designed to solve exact similarity queries.
+
+There exist several packages performing approximate similarity search, like [`Rayuela.jl`](https://github.com/una-dinosauria/Rayuela.jl) using product quantization schemes, the wrapper for the [`FAISS`](https://faiss.ai/) library [`Faiss.jl`](https://github.com/zsz00/Faiss.jl). The FAISS library provides high-performance implementations of product quantization schemes and locality-sensitive hashing schemes, along with an industrial-strength implementation of the [`HNSW`](https://github.com/nmslib/hnswlib) index. The [`NearestNeighborDescent.jl`](https://github.com/dillondaudert/NearestNeighborDescent.jl) implements the search algorithm behind [`pynndescent`](https://pynndescent.readthedocs.io/en/latest/?badge=latest).
+
+The `SimilaritySearch.jl` package tries to enrich the ecosystem with search structures and algorithms designed to take advantage of multithreading systems and a unique autotuning feature that simplifies its usage for practitioners. These features are succinctly and efficiently implemented due to the Julia programming language dynamism and performance.
+Regarding performance characteristics, the construction times are vastly reduced compared to similar approaches without reducing search performance or result quality.
 
 # Installing SimilaritySearch
 
@@ -46,11 +54,11 @@ Contributions are welcome. Please fill a pull request for documentating and impl
 Report issues in the package providing a minimal reproducible example. If the issue is data dependant, please don't forget to provide the necessary data to reproduce it.
 
 ## Limitations of `SearchGraph`
[…]
