### 2022-06-28 15cb708
removes ; from paper.md
