+++ README.md
@@ -88,9 +88,9 @@ The current algorithm (version `0.8` and `0.9`) is described and benchmarked in 
 }
 ```
 
-The package paper is under review of the Journal of Open Source Software (JOSS)
+This package is also described in the JOSS paper, currently under review.
 
-> Eric S. Tellez and Guillermo Ruiz. _`SimilaritySearch.jl`: Autotuned nearest neighbor indexes for Julia_. JOSS (Under review)
+> Eric S. Tellez and Guillermo Ruiz. _`SimilaritySearch.jl`: Autotuned nearest neighbor indexes for Julia_. Journal of Open Source Software (Under review)
 
 ## About v0.9.X series
 
