### 2022-06-29 #4442 @sadit
Thank you again for your time. We decided to update instructions and code to v0.9.3, the latest version. I tried to update the version, but editorialbot indicated that it is an editor-level command.

Please let me discuss changes point by point. Regarding suggestions of @ahwillia:

> Ok, I do have one very small suggestion for @sadit -- it would be nice to add some links at the top of the GitHub README to other alternative libraries in Julia like [`NearestNeighbors.jl`](https://github.com/KristofferC/NearestNeighbors.jl) which seems to also be actively maintained. My understanding is that `SimilaritySearch.jl` is better suited to find approximate solutions on very large datasets than existing options. Many users will not read the pdf compiled for JOSS so it would be nice to briefly summarize these alternative libraries on the main page and how `SimilaritySearch.jl` is different.

Thank you for your valuable comments. We added a brief section in the README.md (_Similarity search *ecosystem* in Julia_) where we discuss packages that solve the same problem and some of their main differences. The new paragraphs are reproduced here:

>>Currently, there exists several packages dedicated to nearest neighbor search, for instance we have [`NearestNeighbors.jl`](https://github.com/KristofferC/NearestNeighbors.jl), [`RegionTrees.jl`](https://github.com/rdeits/RegionTrees.jl), and [`JuliaNeighbors`](https://github.com/JuliaNeighbors) implement search structures like [kd-trees](https://en.wikipedia.org/wiki/K-d_tree), [ball trees](https://en.wikipedia.org/wiki/Ball_tree), [quadtrees](https://en.wikipedia.org/wiki/Quadtree), [octrees](https://en.wikipedia.org/wiki/Octree), [bk-trees](https://en.wikipedia.org/wiki/BK-tree), [vp-tree](https://en.wikipedia.org/wiki/Vantage-point_tree) and other multidimensional and metric structures. These structures work quite well for low dimensional data since they are designed to solve exact similarity queries.

>>There exist several packages performing approximate similarity search, like [`Rayuela.jl`](https://github.com/una-dinosauria/Rayuela.jl) using product quantization schemes, the wrapper for the [`FAISS`](https://faiss.ai/) library [`Faiss.jl`](https://github.com/zsz00/Faiss.jl). The FAISS library provides high-performance implementations of product quantization schemes and locality-sensitive hashing schemes, along with an industrial-strength implementation of the [`HNSW`](https://github.com/nmslib/hnswlib) index. The [`NearestNeighborDescent.jl`](https://github.com/dillondaudert/NearestNeighborDescent.jl) implements the search algorithm behind [`pynndescent`](https://pynndescent.readthedocs.io/en/latest/?badge=latest).

>>The `SimilaritySearch.jl` package tries to enrich the ecosystem with search structures and algorithms designed to take advantage of multithreading systems and a unique autotuning feature that simplifies its usage for practitioners. These features are succinctly and efficiently implemented due to the Julia programming language dynamism and performance.
Regarding performance characteristics, the construction times are vastly reduced compared to similar approaches without reducing search performance or result quality.

Regarding the observations of @kose-y,

> @sadit -- It looks like a great package, and the authors did a thorough job with the submission. However, the example code provided in the paper has a minor error:
>
> Code line 20:
>
> ```julia
> example()
> ```
>
> must have a value of k as an argument.
>
> @jbytecode I have a question about the package version consistency. The submitted version of the package is `v0.8.20`, while the installation instruction would install the newest version, `v0.9.3`. There has been a breaking change between them, so the example code in the paper would not work correctly with the newer version. Specifically, the keyword argument `parallel` has been removed in `v0.9`. Should the authors provide the installation instruction for `v0.8.20` for their installation instruction in the paper?

Thank you for your valuable comments. We fixed the code, adding the missing parameter and working on 0.9. As you mentioned, we only removed `parallel=true` since it is now the default. The extra semicolon was also removed.

The benchmarks were also _rerun_; we decreased running times due to using v0.9.3 since it has several minor speed enhancements. The operations construction, optimization, search, searchbatch, and closestpair improved their speed (perhaps a slight improvement in some cases) without modifying their quality. On the other hand, the `allknn` was simplified, and, as a result, while it is faster, the recall scores were reduced by a few percentual points, as compared with `0.8` runs. Improving `allknn` results are part of our current research goals. Therefore, if we have some luck, some improvements are likely to be introduced in the following versions of `SimilaritySearch.jl.`
