+++ joss/paper.bib
@@ -1,3 +1,15 @@
+@article{bezanson2017julia,
+  title     = {{J}ulia: A fresh approach to numerical computing},
+  author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
+  journal   = {SIAM review},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  year      = {2017},
+  publisher = {SIAM},
+  url       = {https://doi.org/10.1137/141000671}
+}
+
 @misc{simsearch2022,
   doi = {10.48550/ARXIV.2201.07917},
   url = {https://arxiv.org/abs/2201.07917},
