### 2022-06-29 #4442 @jbytecode
@sadit - I have just prepared and send a pull request that includes possible correction in manuscript and bibtex.

Please review it carefully and if you are agreed with them, please apply.

Thank you.
