### 2022-06-29 bb7568d
adds suggestions, fixes some typos, and add acknowledgements
