### 2022-06-29 dd00e4c
adds doi for SIAM Julia's paper
