+++ joss/paper.bib
@@ -7,7 +7,8 @@
   pages     = {65--98},
   year      = {2017},
   publisher = {SIAM},
-  url       = {https://doi.org/10.1137/141000671}
+  url       = {https://doi.org/10.1137/141000671},
+  doi = {10.1137/141000671}
 }
 
 @misc{simsearch2022,
