### 2022-06-29 #4442 @sadit
@jbytecode

The package already has tagged releases since that is how it works with the Julia registry and package manager. Should I create a new release?
