### 2022-06-29 cf7d01d
fixes a couple of typos
