### 2022-06-29 #4442 @sadit
@jbytecode

We reviewed and updated the manuscript (fixed a couple of typos). We also created the Zenodo archive. The requested information is as follows:

- The tagged version is `v0.9.3`
- The zenodo doi is `10.5281/zenodo.6779410`
- The zenodo archive URL is `https://zenodo.org/record/6779410`
