### 2022-06-29 #4442 @jbytecode
@openjournals/dev compilation was success in `generate pdf`, but failed in `acceptance`. could you please help us?
