### 2022-06-29 #4442 @sadit
@jbytecode

I think that the error comes from using a slash as part of the table label. I removed the slash in the label in a new commit
