### 2022-06-30 #4442 @tarleb
The pipeline was updated and should now compile the paper as is.

One note though: while working on this, I noticed that the bibliography entry for Witten et al. (1999) lists one of the authors twice, another even three times. That should probably be fixed.

And last but not least: congrats on the acceptance!
