### 2022-06-30 9473223
fixes managing GB bibentry
