+++ joss/paper.bib
@@ -24,7 +24,7 @@
 
 @book{witten1999managing,
   title={Managing gigabytes: compressing and indexing documents and images},
-  author={Witten, Ian H and Witten, Ian H and Moffat, Alistair and Bell, Timothy C and Bell, Timothy C and Fox, Ed and Bell, Timothy C},
+  author={Witten, Ian H and Moffat, Alistair and Bell, Timothy C},
   year={1999},
   publisher={Morgan Kaufmann}
 }
