### 2022-07-05 #4442 @danielskatz
👋 @sadit - I'm the JOSS AEiC this week handling the rest of the process for your paper.  I've suggested some changes in https://github.com/sadit/SimilaritySearch.jl/pull/23 - please merge this or let me know what you disagree with.
