### 2022-07-05 31e65c7
two more small changes
